requirements.txt
# 核心依赖
crewai>=0.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# API客户端
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0
httpx>=0.24.0

# 数据处理
pymongo>=4.5.0
ormsgpack>=1.3.0

# 媒体处理
opencv-python>=4.8.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# 工具
numpy>=1.24.0
pillow>=10.0.0